feat: UpdateClassSchedules handle classroom conflict validation

+++ Back/Features/Classes/UpdateClassSchedules/UpdateClassSchedulesService.cs
@@ -27,6 +27,9 @@ public class UpdateClassSchedulesService(EstudDbContext ctx) : IEstudService
         var teacherSchedulesResult = await ValidateTeacherSchedules(institutionId, classId, teacherIds, schedules);
         if (teacherSchedulesResult.IsError) return teacherSchedulesResult.Error;
 
+        var classroomSchedulesResult = await ValidateClassroomSchedules(institutionId, classId, schedules);
+        if (classroomSchedulesResult.IsError) return classroomSchedulesResult.Error;
+
         ctx.Schedules.RemoveRange(@class.Schedules);
         @class.Schedules = schedules;
         await ctx.SaveChangesAsync();
@@ -73,6 +76,52 @@ public class UpdateClassSchedulesService(EstudDbContext ctx) : IEstudService
             var otherSchedules = teacherSchedules.Where(x => x.TeacherId == teacherId).ToList();
 
             if (slotsForTeacher.Any(x => otherSchedules.Any(x.Conflict))) return TeacherScheduleConflict.I;
+        }
+
+        return EstudSuccess.I;
+    }
+
+    private async Task<OneOf<EstudSuccess, EstudError>> ValidateClassroomSchedules(
+        int institutionId,
+        int classId,
+        List<Schedule> schedules)
+    {
+        var classroomIds = schedules.Where(s => s.ClassroomId != null).Select(s => s.ClassroomId!.Value).Distinct().ToList();
+        if (classroomIds.Count == 0) return EstudSuccess.I;
+
+        var classroomsCount = await ctx.Classrooms.CountAsync(c => c.InstitutionId == institutionId && classroomIds.Contains(c.Id));
+        if (classroomsCount != classroomIds.Count) return ClassroomNotFound.I;
+
+        const string sql = @"
+            SELECT
+                s.*
+            FROM
+                estud.classes c
+            INNER JOIN
+                estud.schedules s ON s.class_id = c.id
+            WHERE
+                c.institution_id = {0}
+                    AND
+                c.id <> {1}
+                    AND
+                c.status <> {2}
+                    AND
+                s.classroom_id = ANY({3})
+            GROUP BY
+                s.id
+        ";
+        var classroomSchedules = await ctx.Schedules
+            .FromSqlRaw(sql, institutionId, classId, ClassStatus.Finalized.ToInt(), classroomIds.ToArray())
+            .AsNoTracking().ToListAsync();
+
+        // Conflito de agenda por sala
+        foreach (var classroomId in classroomIds)
+        {
+            var slotsForClassroom = schedules.Where(s => s.ClassroomId == classroomId).ToList();
+
+            var otherSchedules = classroomSchedules.Where(x => x.ClassroomId == classroomId).ToList();
+
+            if (slotsForClassroom.Any(x => otherSchedules.Any(x.Conflict))) return ClassroomScheduleConflict.I;
         }
 
         return EstudSuccess.I;
